+++ src/main/java/coupon/coupons/repository/CouponRepository.java
@@ -1,7 +1,7 @@
-package coupon.repository;
+package coupon.coupons.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import coupon.domain.Coupon;
+import coupon.coupons.domain.Coupon;
 
 public interface CouponRepository extends JpaRepository<Coupon, Long> {
 }
