+++ src/main/java/coupon/coupons/service/CouponService.java
@@ -1,11 +1,11 @@
-package coupon.service;
+package coupon.coupons.service;
 
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import coupon.domain.Coupon;
-import coupon.repository.CouponRepository;
+import coupon.coupons.domain.Coupon;
+import coupon.coupons.repository.CouponRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
