+++ src/main/java/coupon/member/domain/Member.java
@@ -1,4 +1,4 @@
-package member.domain;
+package coupon.member.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
