+++ src/test/java/coupon/coupons/domain/CategoryTest.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.coupons.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
