+++ src/test/java/coupon/coupons/domain/MemberCouponTest.java
@@ -1,10 +1,10 @@
-package coupon.domain;
+package coupon.coupons.domain;
 
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import member.domain.Member;
-import membercoupon.domain.MemberCoupon;
+import coupon.member.domain.Member;
+import coupon.membercoupon.domain.MemberCoupon;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
